refactor(metadata_store): converge FS/S3 onto angos-storage with unified Coordinator
