+++ CHANGELOG.md
@@ -46,7 +46,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Outbound `[registry_client]` blocks with a partial mTLS pair (`client_certificate` set without `client_private_key`, or vice versa) now fail at configuration load instead of silently disabling outbound client authentication.
 - Webhook retries are capped at 16 with saturating backoff arithmetic.
 - The S3 shard-key hash uses SHA-256, so shard placement is stable across processes and builds.
-- Filesystem metadata now uses the same sharded blob-index and namespace-registry layout as S3, and scrub migrates legacy filesystem and S3 metadata files to the sharded layout.
+- Filesystem metadata now uses the same sharded blob-index and namespace-registry layout as S3 for newly written data. Pre-existing legacy `index.json` and `namespace_registry.json` files keep working: runtime reads consult the sharded layout first and fall back to the legacy file, and writes are applied in place to a legacy file when one is present so the layout does not split mid-blob. There is no auto-migration on the hot path; only `scrub` (and only `scrub`) converts legacy data into the sharded layout and deletes the legacy file afterwards.
 - The Redis client reuses a multiplexed connection across operations and backs off on lock contention.
 - The S3 backend now uses a custom HTTP client in place of the AWS SDK: focused retry / timeout / signing tailored to the operations the registry actually issues, smaller binary, and no AWS SDK transitive dependencies.
 - Filesystem cleanup of empty ancestor directories is depth-bounded (2 levels for blob shards, 3 for index and namespace-registry shards, 4 for link containers) and rooted-subtree guarded; a misshaped path can never walk above the configured root.
