+++ src/registry/metadata_store/backend/namespace_registry.rs
@@ -1,0 +1,282 @@
+use bytes::Bytes;
+use futures_util::stream::{self, StreamExt};
+use tracing::{debug, warn};
+
+use angos_storage::{ConditionalStore, Error as StorageError};
+
+use crate::registry::{
+    metadata_store::{
+        Error,
+        backend::{
+            Backend,
+            coordinator::{MAX_CAS_RETRIES, sleep_cas_retry},
+        },
+        lock_ops::with_validated_lock,
+        sharded::{self, NamespaceRegistry},
+    },
+    path_builder,
+};
+
+impl Backend {
+    pub async fn read_namespace_registry(&self) -> Result<Option<NamespaceRegistry>, Error> {
+        let shard_dir = path_builder::namespace_registry_shard_dir();
+        let mut all_namespaces = Vec::new();
+        let mut found_shards = false;
+        let mut token = None;
+
+        loop {
+            let page = self
+                .store()
+                .list_children(&shard_dir, 1000, token, None)
+                .await?;
+
+            if !page.objects.is_empty() {
+                found_shards = true;
+            }
+
+            let shard_results: Vec<Result<Vec<String>, Error>> =
+                stream::iter(page.objects.into_iter().map(|obj| {
+                    let shard_path = format!("{shard_dir}/{obj}");
+                    async move {
+                        match self.store().get(&shard_path).await {
+                            Ok(data) => match serde_json::from_slice::<NamespaceRegistry>(&data) {
+                                Ok(registry) => Ok(registry.namespaces),
+                                Err(_) => Ok(Vec::new()),
+                            },
+                            Err(StorageError::NotFound) => Ok(Vec::new()),
+                            Err(e) => Err(Error::from(e)),
+                        }
+                    }
+                }))
+                .buffer_unordered(sharded::SHARD_READ_CONCURRENCY)
+                .collect()
+                .await;
+
+            for result in shard_results {
+                all_namespaces.extend(result?);
+            }
+
+            token = page.next_token;
+            if token.is_none() {
+                break;
+            }
+        }
+
+        if !found_shards {
+            let path = path_builder::namespace_registry_path();
+            return match self.store().get(&path).await {
+                Ok(data) => match serde_json::from_slice(&data) {
+                    Ok(registry) => Ok(Some(registry)),
+                    Err(e) => {
+                        warn!("Corrupt namespace registry, will rebuild: {e}");
+                        Ok(None)
+                    }
+                },
+                Err(StorageError::NotFound) => Ok(None),
+                Err(e) => Err(e.into()),
+            };
+        }
+
+        let all_namespaces = sharded::normalize_namespaces(all_namespaces);
+        Ok(Some(NamespaceRegistry {
+            namespaces: all_namespaces,
+        }))
+    }
+
+    pub async fn rebuild_namespace_registry(&self) -> Result<(), Error> {
+        let repo_dir = path_builder::repository_dir();
+        let namespaces = self.collect_namespaces(repo_dir, "").await?;
+        let namespaces = sharded::normalize_namespaces(namespaces);
+        let shards = sharded::group_namespaces_by_shard(namespaces);
+        for (shard_key, shard_namespaces) in &shards {
+            self.write_namespace_shard(shard_key, shard_namespaces)
+                .await?;
+        }
+        Ok(())
+    }
+
+    async fn write_namespace_shard(
+        &self,
+        shard_key: &str,
+        shard_namespaces: &[String],
+    ) -> Result<(), Error> {
+        let registry = sharded::registry_for_namespaces(shard_namespaces);
+        let content = Bytes::from(serde_json::to_vec(&registry)?);
+        let path = format!(
+            "{}/{shard_key}.json",
+            path_builder::namespace_registry_shard_dir()
+        );
+
+        if let Some(cs) = self.coordinator.conditional_store() {
+            cas_write_namespace_shard(cs, &path, content).await
+        } else {
+            let lock_key = format!("namespace_registry_shard_{shard_key}");
+            let lock_keys = [lock_key];
+            with_validated_lock(
+                self.lock(),
+                &lock_keys,
+                "lock invalidated during namespace registry rebuild",
+                move || async move { self.store().put(&path, content).await.map_err(Error::from) },
+            )
+            .await
+        }
+    }
+
+    pub async fn register_namespace(&self, namespace: &str) -> Result<(), Error> {
+        if self.known_namespaces.lock().await.contains(namespace) {
+            return Ok(());
+        }
+
+        let shard_key = path_builder::namespace_shard_key(namespace);
+        let path = path_builder::namespace_registry_shard_path(namespace);
+
+        if let Some(cs) = self.coordinator.conditional_store() {
+            cas_register_namespace(cs, namespace, &path).await?;
+        } else {
+            let lock_key = format!("namespace_registry_shard_{shard_key}");
+            let lock_keys = [lock_key];
+            with_validated_lock(
+                self.lock(),
+                &lock_keys,
+                "lock invalidated during namespace registry update",
+                || async {
+                    let mut registry = match self.store().get(&path).await {
+                        Ok(data) => {
+                            serde_json::from_slice::<NamespaceRegistry>(&data).unwrap_or_default()
+                        }
+                        Err(StorageError::NotFound) => NamespaceRegistry::default(),
+                        Err(e) => return Err(Error::from(e)),
+                    };
+                    if sharded::insert_sorted_unique(&mut registry.namespaces, namespace) {
+                        let content = Bytes::from(serde_json::to_vec(&registry)?);
+                        self.store().put(&path, content).await?;
+                    }
+                    Ok(())
+                },
+            )
+            .await?;
+        }
+
+        self.known_namespaces
+            .lock()
+            .await
+            .insert(namespace.to_string());
+        Ok(())
+    }
+
+    pub async fn collect_namespaces(
+        &self,
+        root_path: &str,
+        root_prefix: &str,
+    ) -> Result<Vec<String>, Error> {
+        let mut stack: Vec<(String, String)> =
+            vec![(root_path.to_string(), root_prefix.to_string())];
+        let mut namespaces = Vec::new();
+
+        while let Some((path, prefix)) = stack.pop() {
+            let mut token = None;
+            loop {
+                let page = self.store().list_children(&path, 1000, token, None).await?;
+
+                let mut children = Vec::new();
+                for entry in &page.sub_prefixes {
+                    if entry.starts_with('_') {
+                        let namespace = prefix.strip_suffix('/').unwrap_or(&prefix);
+                        if !namespace.is_empty() {
+                            namespaces.push(namespace.to_string());
+                        }
+                        continue;
+                    }
+                    let child_path = format!("{path}/{entry}");
+                    let child_prefix = format!("{prefix}{entry}/");
+                    children.push((child_path, child_prefix));
+                }
+                for child in children.into_iter().rev() {
+                    stack.push(child);
+                }
+
+                token = page.next_token;
+                if token.is_none() {
+                    break;
+                }
+            }
+        }
+
+        Ok(namespaces)
+    }
+}
+
+// ───────────────────────────────────────────────────────────────────────────
+// CAS helpers (free functions to avoid borrow-checker issues with async)
+// ───────────────────────────────────────────────────────────────────────────
+
+async fn cas_write_namespace_shard(
+    cs: &dyn ConditionalStore,
+    path: &str,
+    content: Bytes,
+) -> Result<(), Error> {
+    let etag = match cs.get_with_etag(path).await {
+        Ok((_, etag)) => etag,
+        Err(StorageError::NotFound) => None,
+        Err(e) => return Err(e.into()),
+    };
+    let write_result = if let Some(ref etag) = etag {
+        cs.put_if_match(path, etag, content).await.map(|_| ())
+    } else {
+        cs.put_if_absent(path, content).await.map(|_| ())
+    };
+    match write_result {
+        Ok(()) => {}
+        Err(StorageError::PreconditionFailed) => {
+            debug!("Namespace registry shard rebuild lost CAS race; concurrent write wins");
+        }
+        Err(e) => return Err(e.into()),
+    }
+    Ok(())
+}
+
+async fn cas_register_namespace(
+    cs: &dyn ConditionalStore,
+    namespace: &str,
+    path: &str,
+) -> Result<(), Error> {
+    for attempt in 0..MAX_CAS_RETRIES {
+        let (mut registry, etag) = match cs.get_with_etag(path).await {
+            Ok((data, etag)) => (
+                serde_json::from_slice::<NamespaceRegistry>(&data).unwrap_or_default(),
+                etag,
+            ),
+            Err(StorageError::NotFound) => (NamespaceRegistry::default(), None),
+            Err(e) => return Err(e.into()),
+        };
+
+        if !sharded::insert_sorted_unique(&mut registry.namespaces, namespace) {
+            return Ok(());
+        }
+
+        let content = Bytes::from(serde_json::to_vec(&registry)?);
+        let write_result = if let Some(ref etag) = etag {
+            cs.put_if_match(path, etag, content).await.map(|_| ())
+        } else {
+            cs.put_if_absent(path, content).await.map(|_| ())
+        };
+
+        match write_result {
+            Ok(()) => return Ok(()),
+            Err(StorageError::PreconditionFailed) => {
+                debug!(
+                    namespace,
+                    attempt, "Namespace registry shard CAS conflict, retrying"
+                );
+                sleep_cas_retry(attempt).await;
+            }
+            Err(e) => return Err(e.into()),
+        }
+    }
+
+    warn!(
+        namespace,
+        "Namespace registry shard CAS retries exhausted; namespace will appear after scrub reconciles"
+    );
+    Ok(())
+}
