+++ src/registry/metadata_store/config.rs
@@ -6,12 +6,13 @@ use tracing::info;
 use crate::{
     cache::Cache,
     registry::{
-        blob_store, metadata_store,
-        metadata_store::{ConditionalCapabilities, Error, LockStrategy, MetadataStore},
+        blob_store,
+        metadata_store::{ConditionalCapabilities, Error, LockStrategy, MetadataStore, fs, s3},
     },
     secret::Secret,
 };
 use angos_s3_client as s3_client;
+use angos_storage::s3::Backend as StorageS3Backend;
 
 #[derive(Clone, Debug, Default, Deserialize, PartialEq)]
 #[allow(clippy::large_enum_variant)]
@@ -24,16 +25,16 @@ pub enum MetadataStoreConfig {
     #[default]
     Inherit,
     #[serde(rename = "fs")]
-    FS(metadata_store::fs::BackendConfig),
+    FS(fs::BackendConfig),
     #[serde(rename = "s3")]
-    S3(metadata_store::s3::BackendConfig),
+    S3(s3::BackendConfig),
 }
 
 impl MetadataStoreConfig {
     pub fn from_blob_store(blob: &blob_store::BlobStorageConfig) -> Self {
         match blob {
             blob_store::BlobStorageConfig::FS(config) => {
-                MetadataStoreConfig::FS(metadata_store::fs::BackendConfig {
+                MetadataStoreConfig::FS(fs::BackendConfig {
                     root_dir: config.root_dir.clone(),
                     sync_to_disk: config.sync_to_disk,
                     ..Default::default()
@@ -41,7 +42,7 @@ impl MetadataStoreConfig {
             }
             blob_store::BlobStorageConfig::S3(config) => {
                 info!("Auto-configuring S3 metadata-store from blob-store");
-                MetadataStoreConfig::S3(metadata_store::s3::BackendConfig {
+                MetadataStoreConfig::S3(s3::BackendConfig {
                     bucket: config.bucket.clone(),
                     region: config.region.clone(),
                     endpoint: config.endpoint.clone(),
@@ -61,9 +62,15 @@ impl MetadataStoreConfig {
                  Configuration::resolve_metadata_config before probe"
             ),
             MetadataStoreConfig::S3(config) => {
-                let store = s3_client::Backend::new(&config.to_data_store_config())
+                let http = s3_client::Backend::new(&config.to_data_store_config())
                     .map_err(|e| Error::StorageBackend(e.to_string()))?;
-                let caps = metadata_store::s3::probe_conditional_capabilities(&store).await?;
+                let storage = Arc::new(
+                    StorageS3Backend::builder()
+                        .client(Arc::new(http))
+                        .build()
+                        .map_err(|e| Error::StorageBackend(e.to_string()))?,
+                );
+                let caps = s3::probe_conditional_capabilities(storage.as_ref()).await?;
                 if matches!(config.lock_strategy, LockStrategy::S3(_)) && !caps.supports_cas() {
                     return Err(Error::Lock(format!(
                         "S3 lock strategy requires If-None-Match and If-Match support, \
@@ -93,9 +100,7 @@ impl MetadataStoreConfig {
                 "MetadataStoreConfig::Inherit must be resolved via \
                  Configuration::resolve_metadata_config before to_backend"
             ),
-            MetadataStoreConfig::FS(config) => {
-                Ok((Arc::new(metadata_store::fs::Backend::new(config)?), None))
-            }
+            MetadataStoreConfig::FS(config) => Ok((Arc::new(config.to_backend()?), None)),
             MetadataStoreConfig::S3(config) => {
                 let caps = match &config.capabilities {
                     Some(declared) => {
@@ -113,11 +118,7 @@ impl MetadataStoreConfig {
                     }
                     None => self.probe().await?,
                 };
-                let backend = metadata_store::s3::Backend::new(config, caps.clone())?;
-                let backend = match cache {
-                    Some(c) => backend.with_cache(c),
-                    None => backend,
-                };
+                let backend = config.to_backend(caps.clone(), cache)?;
                 Ok((Arc::new(backend), caps))
             }
         }
@@ -130,7 +131,7 @@ mod tests {
     use crate::{registry::metadata_store::lock::s3::S3LockConfig, secret::Secret};
 
     fn s3_config_with_lock_strategy(lock_strategy: LockStrategy) -> MetadataStoreConfig {
-        MetadataStoreConfig::S3(metadata_store::s3::BackendConfig {
+        MetadataStoreConfig::S3(s3::BackendConfig {
             access_key_id: Secret::new("root".to_string()),
             secret_key: Secret::new("roottoor".to_string()),
             endpoint: "http://127.0.0.1:9000".to_string(),
@@ -176,7 +177,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_probe_fs_config_is_noop() {
-        let config = MetadataStoreConfig::FS(metadata_store::fs::BackendConfig {
+        let config = MetadataStoreConfig::FS(fs::BackendConfig {
             root_dir: "/tmp/probe-test".to_string(),
             lock_strategy: LockStrategy::Memory,
             sync_to_disk: false,
