--- src/registry/metadata_store/s3/coordinator.rs
@@ -1,626 +1,0 @@
-use std::{collections::HashMap, io::ErrorKind, sync::Arc};
-
-use async_trait::async_trait;
-use bytes::Bytes;
-use futures_util::future::join_all;
-use tracing::{debug, info, instrument, warn};
-
-use super::{Backend, MAX_BLOB_INDEX_CAS_RETRIES, config::BackendConfig, sleep_cas_retry};
-use crate::{
-    oci::Digest,
-    registry::{
-        metadata_store::{
-            BlobIndexOperation, ConditionalCapabilities, Error, LinkMetadata, LinkOperation,
-            LockStrategy, ResolvedCreate, ResolvedDelete,
-            link_kind::LinkKind,
-            lock::{self, LockBackend, LockGuard, MemoryBackend, S3LockBackend},
-            lock_ops::{LockOps, blob_index_lock_key, link_lock_key, with_validated_lock},
-            sharded, update_links,
-        },
-        path_builder,
-    },
-};
-use angos_s3_client as s3_client;
-
-#[async_trait]
-pub trait WriteCoordinator: Send + Sync + std::fmt::Debug {
-    async fn update_links(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        operations: &[LinkOperation],
-    ) -> Result<(), Error>;
-
-    async fn update_blob_index(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        digest: &Digest,
-        operation: BlobIndexOperation,
-    ) -> Result<(), Error>;
-
-    async fn acquire_blob_data_lock(&self, digest: &Digest) -> Result<LockGuard, Error>;
-
-    async fn register_namespace(&self, backend: &Backend, namespace: &str) -> Result<(), Error>;
-
-    async fn rebuild_namespace_registry_shard(
-        &self,
-        backend: &Backend,
-        shard_key: &str,
-        path: &str,
-        content: Bytes,
-    ) -> Result<(), Error>;
-
-    async fn touch_link_access_time(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        link: &LinkKind,
-    ) -> Result<LinkMetadata, Error>;
-
-    async fn flush_access_time(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        link: &LinkKind,
-    ) -> Result<(), Error>;
-}
-
-#[derive(Debug)]
-pub struct CasCoordinator {
-    pub lock: Arc<S3LockBackend>,
-}
-
-#[derive(Debug)]
-pub struct LockCoordinator {
-    pub lock: Arc<dyn LockBackend + Send + Sync>,
-}
-
-impl CasCoordinator {
-    /// Executes link operations under a distributed lock, returning pending
-    /// blob index operations for CAS updates after the lock is released.
-    async fn execute_locked_cas_updates(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        operations: &[LinkOperation],
-    ) -> Result<HashMap<Digest, Vec<BlobIndexOperation>>, Error> {
-        if operations.is_empty() {
-            return Ok(HashMap::new());
-        }
-
-        let mut lock_keys: Vec<String> = operations
-            .iter()
-            .map(|op| {
-                let link = match op {
-                    LinkOperation::Create { link, .. } | LinkOperation::Delete { link, .. } => link,
-                };
-                link_lock_key(namespace, link)
-            })
-            .collect();
-        lock_keys.sort();
-        lock_keys.dedup();
-
-        let guard = self.lock.acquire(&lock_keys).await?;
-
-        let read_results = join_all(operations.iter().map(|op| async move {
-            let link = match op {
-                LinkOperation::Create { link, .. } | LinkOperation::Delete { link, .. } => link,
-            };
-            let metadata = backend.read_link_reference(namespace, link).await.ok();
-            (op, metadata)
-        }))
-        .await;
-
-        let mut link_cache: HashMap<LinkKind, LinkMetadata> = HashMap::new();
-        let mut creates: Vec<ResolvedCreate> = Vec::new();
-        let mut deletes: Vec<ResolvedDelete> = Vec::new();
-
-        for (op, metadata) in &read_results {
-            match op {
-                LinkOperation::Create {
-                    link,
-                    target,
-                    referrer,
-                    media_type,
-                    descriptor,
-                } => {
-                    let old_target = metadata.as_ref().map(|m| m.target.clone());
-                    if let Some(m) = metadata {
-                        link_cache.insert(link.clone(), m.clone());
-                    }
-                    creates.push(ResolvedCreate {
-                        link: link.clone(),
-                        target: target.clone(),
-                        old_target,
-                        referrer: referrer.clone(),
-                        media_type: media_type.clone(),
-                        descriptor: descriptor.as_deref().cloned(),
-                    });
-                }
-                LinkOperation::Delete { link, referrer } => {
-                    if let Some(m) = metadata {
-                        link_cache.insert(link.clone(), m.clone());
-                        deletes.push(ResolvedDelete {
-                            link: link.clone(),
-                            target: m.target.clone(),
-                            referrer: referrer.clone(),
-                        });
-                    }
-                }
-            }
-        }
-
-        if creates.is_empty() && deletes.is_empty() {
-            guard.release().await;
-            return Ok(HashMap::new());
-        }
-
-        if !guard.is_valid() {
-            guard.release().await;
-            return Err(Error::Lock("lock invalidated during operation".into()));
-        }
-
-        let pending_blob_ops = backend
-            .apply_link_operations(namespace, &creates, &deletes, &mut link_cache)
-            .await?;
-
-        guard.release().await;
-
-        Ok(pending_blob_ops)
-    }
-}
-
-#[async_trait]
-impl WriteCoordinator for CasCoordinator {
-    #[instrument(name = "update_links_cas", skip(self, backend, operations))]
-    async fn update_links(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        operations: &[LinkOperation],
-    ) -> Result<(), Error> {
-        let mut locked_ops: Vec<LinkOperation> = Vec::new();
-        let mut any_creates = false;
-
-        let mut optimistic: Vec<(LinkOperation, LinkKind, Digest, LinkMetadata)> = Vec::new();
-
-        for op in operations {
-            match op {
-                LinkOperation::Create {
-                    link,
-                    target,
-                    referrer,
-                    media_type,
-                    descriptor,
-                } => {
-                    any_creates = true;
-                    let mut metadata = LinkMetadata::from_digest(target.clone())
-                        .with_media_type(media_type.clone())
-                        .with_descriptor(descriptor.as_deref().cloned());
-                    if let Some(r) = referrer {
-                        metadata.add_referrer(r.clone());
-                    }
-                    optimistic.push((op.clone(), link.clone(), target.clone(), metadata));
-                }
-                LinkOperation::Delete { .. } => {
-                    locked_ops.push(op.clone());
-                }
-            }
-        }
-
-        let link_results = join_all(optimistic.iter().map(|(_, link, _, metadata)| {
-            backend.write_link_if_not_exists(namespace, link, metadata)
-        }))
-        .await;
-
-        let mut lockfree_blob_ops: HashMap<Digest, Vec<BlobIndexOperation>> = HashMap::new();
-        for ((op, link, target, metadata), result) in optimistic.into_iter().zip(link_results) {
-            match result {
-                Ok(true) => {
-                    lockfree_blob_ops
-                        .entry(target)
-                        .or_default()
-                        .push(BlobIndexOperation::Insert(link.clone()));
-                    backend.cache_put(namespace, &link, &metadata).await;
-                }
-                Ok(false) => {
-                    locked_ops.push(op);
-                }
-                Err(e) => return Err(e),
-            }
-        }
-
-        let locked_blob_ops = self
-            .execute_locked_cas_updates(backend, namespace, &locked_ops)
-            .await?;
-
-        let mut pending_blob_ops = lockfree_blob_ops;
-        for (digest, ops) in locked_blob_ops {
-            pending_blob_ops.entry(digest).or_default().extend(ops);
-        }
-
-        join_all(pending_blob_ops.iter().map(|(digest, ops)| async move {
-            backend.migrate_blob_index_layout(digest).await?;
-            backend.update_blob_index_cas(namespace, digest, ops).await
-        }))
-        .await
-        .into_iter()
-        .collect::<Result<Vec<_>, _>>()?;
-
-        if any_creates && let Err(e) = backend.register_namespace(namespace).await {
-            warn!(namespace, error = %e, "Failed to register namespace");
-        }
-
-        Ok(())
-    }
-
-    #[instrument(name = "update_blob_index_cas", skip(self, backend))]
-    async fn update_blob_index(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        digest: &Digest,
-        operation: BlobIndexOperation,
-    ) -> Result<(), Error> {
-        backend.migrate_blob_index_layout(digest).await?;
-        backend
-            .update_blob_index_cas(namespace, digest, &[operation])
-            .await
-    }
-
-    async fn acquire_blob_data_lock(&self, digest: &Digest) -> Result<LockGuard, Error> {
-        self.lock.acquire(&[format!("blob-data:{digest}")]).await
-    }
-
-    #[instrument(name = "register_namespace_cas", skip(self, backend))]
-    async fn register_namespace(&self, backend: &Backend, namespace: &str) -> Result<(), Error> {
-        let path = path_builder::namespace_registry_shard_path(namespace);
-
-        for attempt in 0..MAX_BLOB_INDEX_CAS_RETRIES {
-            let (mut registry, etag) = match backend.store.read_with_etag(&path).await {
-                Ok((data, etag)) => {
-                    match serde_json::from_slice::<sharded::NamespaceRegistry>(&data) {
-                        Ok(r) => (r, etag),
-                        Err(_) => (sharded::NamespaceRegistry::default(), etag),
-                    }
-                }
-                Err(e) if e.kind() == ErrorKind::NotFound => {
-                    (sharded::NamespaceRegistry::default(), None)
-                }
-                Err(e) => return Err(Error::from(e)),
-            };
-
-            if !sharded::insert_sorted_unique(&mut registry.namespaces, namespace) {
-                backend
-                    .known_namespaces
-                    .lock()
-                    .await
-                    .insert(namespace.to_string());
-                return Ok(());
-            }
-
-            let content = Bytes::from(serde_json::to_vec(&registry)?);
-            let write_result = if let Some(ref etag) = etag {
-                backend
-                    .store
-                    .put_object_if_match(&path, etag, content)
-                    .await
-                    .map(|_| ())
-            } else {
-                backend
-                    .store
-                    .put_object_if_not_exists(&path, content)
-                    .await
-                    .map(|_| ())
-            };
-
-            match write_result {
-                Ok(()) => {
-                    backend
-                        .known_namespaces
-                        .lock()
-                        .await
-                        .insert(namespace.to_string());
-                    return Ok(());
-                }
-                Err(s3_client::Error::PreconditionFailed) => {
-                    debug!(
-                        namespace,
-                        attempt, "Namespace registry shard CAS conflict, retrying"
-                    );
-                    sleep_cas_retry(attempt).await;
-                }
-                Err(e) => return Err(Error::StorageBackend(e.to_string())),
-            }
-        }
-
-        warn!(
-            namespace,
-            "Namespace registry shard CAS retries exhausted; namespace will appear after scrub reconciles"
-        );
-        Ok(())
-    }
-
-    #[instrument(
-        name = "rebuild_namespace_registry_shard_cas",
-        skip(self, backend, content)
-    )]
-    async fn rebuild_namespace_registry_shard(
-        &self,
-        backend: &Backend,
-        shard_key: &str,
-        path: &str,
-        content: Bytes,
-    ) -> Result<(), Error> {
-        let etag = match backend.store.read_with_etag(path).await {
-            Ok((_, etag)) => etag,
-            Err(e) if e.kind() == ErrorKind::NotFound => None,
-            Err(e) => return Err(Error::from(e)),
-        };
-        let write_result = if let Some(ref etag) = etag {
-            backend
-                .store
-                .put_object_if_match(path, etag, content)
-                .await
-                .map(|_| ())
-        } else {
-            backend
-                .store
-                .put_object_if_not_exists(path, content)
-                .await
-                .map(|_| ())
-        };
-        match write_result {
-            Ok(()) => {}
-            Err(s3_client::Error::PreconditionFailed) => {
-                debug!(
-                    shard_key,
-                    "Namespace registry shard rebuild lost CAS race; concurrent write wins"
-                );
-            }
-            Err(e) => return Err(Error::StorageBackend(e.to_string())),
-        }
-        Ok(())
-    }
-
-    #[instrument(name = "touch_link_access_time_cas", skip(self, backend))]
-    async fn touch_link_access_time(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        link: &LinkKind,
-    ) -> Result<LinkMetadata, Error> {
-        let link_data = backend
-            .read_and_spawn_access_time_update(namespace, link)
-            .await?;
-        backend.cache_put(namespace, link, &link_data).await;
-        Ok(link_data)
-    }
-
-    #[instrument(name = "flush_access_time_cas", skip(self, backend))]
-    async fn flush_access_time(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        link: &LinkKind,
-    ) -> Result<(), Error> {
-        let link_path = path_builder::link_path(link, namespace);
-        let (data, etag) = backend.store.read_with_etag(&link_path).await?;
-        let link_data = LinkMetadata::from_bytes(data)?.accessed();
-        if let Some(etag) = etag {
-            let content =
-                serde_json::to_vec(&link_data).map_err(|e| Error::InvalidData(e.to_string()))?;
-            backend
-                .store
-                .put_object_if_match(&link_path, &etag, content)
-                .await?;
-        }
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl WriteCoordinator for LockCoordinator {
-    #[instrument(name = "update_links_locked", skip(self, backend, operations))]
-    async fn update_links(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        operations: &[LinkOperation],
-    ) -> Result<(), Error> {
-        update_links::run_update_links(backend, &*self.lock, namespace, operations).await
-    }
-
-    #[instrument(name = "update_blob_index_locked", skip(self, backend))]
-    async fn update_blob_index(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        digest: &Digest,
-        operation: BlobIndexOperation,
-    ) -> Result<(), Error> {
-        if backend.conditional_capabilities.supports_cas() {
-            backend.migrate_blob_index_layout(digest).await?;
-            return backend
-                .update_blob_index_cas(namespace, digest, &[operation])
-                .await;
-        }
-
-        let lock_keys = [blob_index_lock_key(digest)];
-        with_validated_lock(
-            &*self.lock,
-            &lock_keys,
-            "lock invalidated during blob index update",
-            || async {
-                backend.migrate_blob_index_layout(digest).await?;
-                backend
-                    .update_blob_index_shard(namespace, digest, &[operation])
-                    .await
-            },
-        )
-        .await
-    }
-
-    async fn acquire_blob_data_lock(&self, digest: &Digest) -> Result<LockGuard, Error> {
-        self.lock.acquire(&[format!("blob-data:{digest}")]).await
-    }
-
-    #[instrument(name = "register_namespace_locked", skip(self, backend))]
-    async fn register_namespace(&self, backend: &Backend, namespace: &str) -> Result<(), Error> {
-        let shard_key = path_builder::namespace_shard_key(namespace);
-        let path = path_builder::namespace_registry_shard_path(namespace);
-        let lock_key = format!("namespace_registry_shard_{shard_key}");
-        let lock_keys = [lock_key];
-
-        with_validated_lock(
-            &*self.lock,
-            &lock_keys,
-            "lock invalidated during namespace registry update",
-            || async {
-                let mut registry = match backend.store.read(&path).await {
-                    Ok(data) => serde_json::from_slice::<sharded::NamespaceRegistry>(&data)
-                        .unwrap_or_default(),
-                    Err(e) if e.kind() == ErrorKind::NotFound => {
-                        sharded::NamespaceRegistry::default()
-                    }
-                    Err(e) => return Err(Error::from(e)),
-                };
-
-                if sharded::insert_sorted_unique(&mut registry.namespaces, namespace) {
-                    let content = Bytes::from(serde_json::to_vec(&registry)?);
-                    backend.store.put_object(&path, content).await?;
-                }
-                Ok(())
-            },
-        )
-        .await?;
-
-        backend
-            .known_namespaces
-            .lock()
-            .await
-            .insert(namespace.to_string());
-        Ok(())
-    }
-
-    #[instrument(
-        name = "rebuild_namespace_registry_shard_locked",
-        skip(self, backend, content)
-    )]
-    async fn rebuild_namespace_registry_shard(
-        &self,
-        backend: &Backend,
-        shard_key: &str,
-        path: &str,
-        content: Bytes,
-    ) -> Result<(), Error> {
-        let lock_key = format!("namespace_registry_shard_{shard_key}");
-        let lock_keys = [lock_key];
-        with_validated_lock(
-            &*self.lock,
-            &lock_keys,
-            "lock invalidated during namespace registry rebuild",
-            move || async move {
-                backend.store.put_object(path, content).await?;
-                Ok(())
-            },
-        )
-        .await
-    }
-
-    #[instrument(name = "touch_link_access_time_locked", skip(self, backend))]
-    async fn touch_link_access_time(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        link: &LinkKind,
-    ) -> Result<LinkMetadata, Error> {
-        let lock_keys = [link_lock_key(namespace, link)];
-        let link_data = with_validated_lock(
-            &*self.lock,
-            &lock_keys,
-            "lock invalidated during access time update",
-            || async {
-                let link_data = backend
-                    .read_link_reference(namespace, link)
-                    .await?
-                    .accessed();
-                backend
-                    .write_link_reference(namespace, link, &link_data)
-                    .await?;
-                Ok(link_data)
-            },
-        )
-        .await?;
-        backend.cache_put(namespace, link, &link_data).await;
-        Ok(link_data)
-    }
-
-    #[instrument(name = "flush_access_time_locked", skip(self, backend))]
-    async fn flush_access_time(
-        &self,
-        backend: &Backend,
-        namespace: &str,
-        link: &LinkKind,
-    ) -> Result<(), Error> {
-        let lock_keys = [link_lock_key(namespace, link)];
-        with_validated_lock(
-            &*self.lock,
-            &lock_keys,
-            "lock invalidated during access time flush",
-            || async {
-                let link_data = backend
-                    .read_link_reference(namespace, link)
-                    .await?
-                    .accessed();
-                backend
-                    .write_link_reference(namespace, link, &link_data)
-                    .await
-            },
-        )
-        .await
-    }
-}
-
-pub(super) fn build_coordinator(
-    config: &BackendConfig,
-    caps: &ConditionalCapabilities,
-) -> Result<Arc<dyn WriteCoordinator>, Error> {
-    match &config.lock_strategy {
-        LockStrategy::S3(s3_lock_config) => {
-            if !caps.supports_cas() {
-                return Err(Error::Lock(format!(
-                    "S3 lock strategy requires If-None-Match and If-Match support, \
-                     but provider has put_if_none_match={}, put_if_match={}. \
-                     Use lock_strategy = redis or lock_strategy = memory instead.",
-                    caps.put_if_none_match, caps.put_if_match
-                )));
-            }
-            info!("Using CAS coordinator with S3 lock for S3 metadata-store");
-            let lock_store = Arc::new(
-                s3_client::Backend::new(&config.to_lock_store_config(s3_lock_config))
-                    .map_err(|e| Error::Lock(format!("Failed to initialize S3 lock store: {e}")))?,
-            );
-            let lock = Arc::new(
-                S3LockBackend::new(lock_store, s3_lock_config, caps.delete_if_match)
-                    .map_err(|e| Error::Lock(format!("Failed to initialize S3 lock store: {e}")))?,
-            );
-            Ok(Arc::new(CasCoordinator { lock }))
-        }
-        LockStrategy::Redis(redis_config) => {
-            info!("Using Redis lock store for S3 metadata-store");
-            let backend = lock::RedisBackend::new(redis_config)
-                .map_err(|e| Error::Lock(format!("Failed to initialize Redis lock store: {e}")))?;
-            Ok(Arc::new(LockCoordinator {
-                lock: Arc::new(backend),
-            }))
-        }
-        LockStrategy::Memory => {
-            info!("Using in-memory lock store for S3 metadata-store");
-            Ok(Arc::new(LockCoordinator {
-                lock: Arc::new(MemoryBackend::new()),
-            }))
-        }
-    }
-}
