+++ src/registry/metadata_store/s3/probe.rs
@@ -1,16 +1,14 @@
-use std::io::ErrorKind;
-
+use bytes::Bytes;
 use tracing::{info, warn};
 
 use crate::registry::metadata_store::{ConditionalCapabilities, Error};
-use angos_s3_client as s3_client;
+use angos_storage::{ConditionalStore, Error as StorageError, Etag};
 
 /// RAII guard that warns if dropped while still armed.
 ///
 /// Create after a probe object is successfully written; call `disarm()` after
-/// explicit cleanup succeeds.  If the function exits early (panic or future
-/// `?` propagation) while the guard is still armed, `Drop` emits a `warn!` so
-/// the orphaned key surfaces in logs.
+/// explicit cleanup succeeds. If the function exits early while the guard is
+/// still armed, `Drop` emits a `warn!` so the orphaned key surfaces in logs.
 struct ProbeGuard {
     key: Option<String>,
 }
@@ -20,7 +18,6 @@ impl ProbeGuard {
         Self { key: Some(key) }
     }
 
-    /// Disarm the guard after successful cleanup so `Drop` stays silent.
     fn disarm(&mut self) {
         self.key.take();
     }
@@ -42,10 +39,10 @@ impl Drop for ProbeGuard {
 /// `DeleteObject If-Match: <etag>` in sequence. Each probe is self-validating:
 /// bogus-ETag attempts verify that the provider actually enforces the condition.
 ///
-/// Returns the probed capabilities. The caller is responsible for failing startup
-/// if any required capability is absent.
+/// Returns the probed capabilities. The caller is responsible for failing
+/// startup if any required capability is absent.
 pub async fn probe_conditional_capabilities(
-    store: &s3_client::Backend,
+    store: &impl ConditionalStore,
 ) -> Result<ConditionalCapabilities, Error> {
     let probe_key = format!("_angos_probe_{}", uuid::Uuid::new_v4());
     probe_conditional_capabilities_with_key(store, &probe_key).await
@@ -53,25 +50,29 @@ pub async fn probe_conditional_capabilities(
 
 /// Inner implementation that accepts an explicit probe key.
 ///
-/// Exposed as `pub(super)` so tests can pass a known key and verify cleanup
+/// Exposed as `pub` so tests can pass a known key and verify cleanup
 /// without having to discover the UUID-suffixed key after the fact.
-pub(super) async fn probe_conditional_capabilities_with_key(
-    store: &s3_client::Backend,
+pub async fn probe_conditional_capabilities_with_key(
+    store: &impl ConditionalStore,
     probe_key: &str,
 ) -> Result<ConditionalCapabilities, Error> {
-    let content: &[u8] = b"probe";
-
-    store.put_object(probe_key, content).await.map_err(|e| {
-        Error::Lock(format!(
-            "conditional capability probe: failed to create probe object: {e}"
-        ))
-    })?;
+    store
+        .put(probe_key, Bytes::from_static(b"probe"))
+        .await
+        .map_err(|e| {
+            Error::Lock(format!(
+                "conditional capability probe: failed to create probe object: {e}"
+            ))
+        })?;
 
     let mut guard = ProbeGuard::new(probe_key.to_string());
 
-    // Test If-None-Match: * — expect 412 because the object already exists.
-    let put_if_none_match = match store.put_object_if_not_exists(probe_key, content).await {
-        Err(s3_client::Error::PreconditionFailed) => true,
+    // Test If-None-Match: * — expect PreconditionFailed because the object already exists.
+    let put_if_none_match = match store
+        .put_if_absent(probe_key, Bytes::from_static(b"probe"))
+        .await
+    {
+        Err(StorageError::PreconditionFailed) => true,
         Ok(_) => {
             warn!(
                 "conditional probe: If-None-Match: * was accepted on existing key; provider does not enforce it"
@@ -85,17 +86,21 @@ pub(super) async fn probe_conditional_capabilities_with_key(
     };
 
     // Test If-Match: <etag> — correct ETag must succeed; bogus ETag must fail.
-    let put_if_match = match store.read_with_etag(probe_key).await {
+    let put_if_match = match store.get_with_etag(probe_key).await {
         Ok((_, Some(etag))) => {
             let correct = store
-                .put_object_if_match(probe_key, &etag, b"updated".to_vec())
+                .put_if_match(probe_key, &etag, Bytes::from_static(b"updated"))
                 .await
                 .is_ok();
             let bogus_rejected = matches!(
                 store
-                    .put_object_if_match(probe_key, "\"bogus\"", b"fail".to_vec())
+                    .put_if_match(
+                        probe_key,
+                        &Etag::new("\"bogus\"".to_string()),
+                        Bytes::from_static(b"fail")
+                    )
                     .await,
-                Err(s3_client::Error::PreconditionFailed)
+                Err(StorageError::PreconditionFailed)
             );
             correct && bogus_rejected
         }
@@ -109,15 +114,14 @@ pub(super) async fn probe_conditional_capabilities_with_key(
         }
     };
 
-    // Test DeleteObject If-Match: <etag> — bogus ETag must fail; correct ETag must succeed.
-    // Re-read the current ETag after the put_if_match update may have changed it.
-    let delete_if_match = match store.read_with_etag(probe_key).await {
+    // Test DeleteObject If-Match: <etag>.
+    let delete_if_match = match store.get_with_etag(probe_key).await {
         Ok((_, Some(etag))) => {
-            // Bogus-ETag attempt first: if the provider ignores the condition and deletes
-            // the object, the correct-ETag attempt below will hit NotFound and correct=false.
             let bogus_rejected = matches!(
-                store.delete_if_match(probe_key, "\"bogus\"").await,
-                Err(s3_client::Error::PreconditionFailed)
+                store
+                    .delete_if_match(probe_key, &Etag::new("\"bogus\"".to_string()))
+                    .await,
+                Err(StorageError::PreconditionFailed)
             );
             let correct = store.delete_if_match(probe_key, &etag).await.is_ok();
             bogus_rejected && correct
@@ -128,7 +132,7 @@ pub(super) async fn probe_conditional_capabilities_with_key(
             );
             false
         }
-        Err(e) if e.kind() == ErrorKind::NotFound => false,
+        Err(StorageError::NotFound) => false,
         Err(e) => {
             warn!("conditional probe: failed to read probe object for delete_if_match test: {e}");
             false
@@ -137,7 +141,7 @@ pub(super) async fn probe_conditional_capabilities_with_key(
 
     // Cleanup — may already have been deleted by the delete_if_match test.
     if let Err(e) = store.delete(probe_key).await
-        && e.kind() != ErrorKind::NotFound
+        && !matches!(e, StorageError::NotFound)
     {
         warn!("conditional probe: cleanup failed for probe object {probe_key}: {e}");
     }
@@ -157,4 +161,135 @@ pub(super) async fn probe_conditional_capabilities_with_key(
     );
 
     Ok(capabilities)
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use crate::{
+        metrics_provider,
+        registry::metadata_store::{
+            LockStrategy,
+            s3::{BackendConfig, probe_conditional_capabilities},
+        },
+        secret::Secret,
+    };
+    use angos_s3_client as s3_client;
+    use angos_storage::{ConditionalStore, Error as StorageError, s3::Backend as StorageS3Backend};
+
+    use crate::registry::metadata_store::s3::probe::probe_conditional_capabilities_with_key;
+
+    fn test_config() -> BackendConfig {
+        metrics_provider::init_for_tests();
+        BackendConfig {
+            access_key_id: Secret::new("root".to_string()),
+            secret_key: Secret::new("roottoor".to_string()),
+            endpoint: "http://127.0.0.1:9000".to_string(),
+            region: "region".to_string(),
+            bucket: "registry".to_string(),
+            key_prefix: format!("test-probe-{}", uuid::Uuid::new_v4()),
+            lock_strategy: LockStrategy::Memory,
+            link_cache_ttl: 0,
+            access_time_debounce_secs: 0,
+            capabilities: None,
+        }
+    }
+
+    fn storage_for(config: &BackendConfig) -> StorageS3Backend {
+        let http = s3_client::Backend::new(&config.to_data_store_config()).unwrap();
+        StorageS3Backend::builder()
+            .client(Arc::new(http))
+            .build()
+            .unwrap()
+    }
+
+    fn storage_for_bucket(config: &BackendConfig, bucket: &str) -> StorageS3Backend {
+        let mut cfg = config.to_data_store_config();
+        cfg.bucket = bucket.to_string();
+        let http = s3_client::Backend::new(&cfg).unwrap();
+        StorageS3Backend::builder()
+            .client(Arc::new(http))
+            .build()
+            .unwrap()
+    }
+
+    #[tokio::test]
+    async fn test_probe_conditional_capabilities() {
+        let config = test_config();
+        let storage = storage_for(&config);
+
+        let result = probe_conditional_capabilities(&storage).await;
+        assert!(result.is_ok(), "Probe should pass on MinIO: {result:?}");
+        let caps = result.unwrap();
+        assert!(caps.put_if_none_match, "MinIO should support If-None-Match");
+        assert!(caps.put_if_match, "MinIO should support If-Match");
+        // MinIO does not enforce If-Match on DeleteObject (returns 200 regardless of ETag),
+        // so delete_if_match is expected to be false.
+        assert!(
+            !caps.delete_if_match,
+            "MinIO does not support conditional delete (ignores If-Match on DeleteObject)"
+        );
+    }
+
+    /// After a successful probe the temporary object must not remain in the bucket.
+    /// Uses a deterministic probe key so the cleanup can be confirmed via a direct
+    /// `get_with_etag` call without having to list the entire bucket.
+    #[tokio::test]
+    async fn test_probe_cleanup_removes_probe_object() {
+        let config = test_config();
+        let storage = storage_for(&config);
+
+        let probe_key = format!("_angos_probe_cleanup_{}", uuid::Uuid::new_v4());
+
+        let result = probe_conditional_capabilities_with_key(&storage, &probe_key).await;
+        assert!(
+            result.is_ok(),
+            "Probe should succeed against MinIO: {result:?}"
+        );
+
+        // The probe object must be absent after the call returns.
+        let read_result = storage.get_with_etag(&probe_key).await;
+        assert!(
+            matches!(read_result, Err(StorageError::NotFound)),
+            "Probe object '{probe_key}' should have been deleted after probe, got: {read_result:?}"
+        );
+    }
+
+    /// Running the probe twice against the same bucket must succeed both times.
+    /// This verifies that successful cleanup in the first run does not leave state
+    /// that would cause the second run to fail (e.g. a leftover object that
+    /// triggers an unexpected If-None-Match success on a fresh key).
+    #[tokio::test]
+    async fn test_probe_idempotent_back_to_back() {
+        let config = test_config();
+        let storage = storage_for(&config);
+
+        let first = probe_conditional_capabilities(&storage).await;
+        assert!(first.is_ok(), "First probe should succeed: {first:?}");
+
+        let second = probe_conditional_capabilities(&storage).await;
+        assert!(second.is_ok(), "Second probe should succeed: {second:?}");
+
+        assert_eq!(
+            first.unwrap(),
+            second.unwrap(),
+            "Probe must return consistent capabilities across successive invocations"
+        );
+    }
+
+    /// When the configured bucket does not exist the probe must return an error
+    /// rather than panicking or hanging, exercising the early-return branch where
+    /// probe-object creation itself fails.
+    #[tokio::test]
+    async fn test_probe_with_nonexistent_bucket_returns_err() {
+        let config = test_config();
+        let storage = storage_for_bucket(&config, "angos-probe-nonexistent-bucket-xyzzy");
+
+        let result = probe_conditional_capabilities(&storage).await;
+        assert!(
+            result.is_err(),
+            "Probe against a non-existent bucket must return Err, got: {result:?}"
+        );
+    }
 }
