+++ src/registry/test_utils.rs
@@ -141,7 +141,7 @@ pub fn backends() -> Vec<Box<dyn RegistryTestCase>> {
 
 pub struct FSRegistryTestCase {
     blob_store: Arc<blob_store::fs::Backend>,
-    metadata_store: Arc<metadata_store::fs::Backend>,
+    metadata_store: Arc<metadata_store::Backend>,
     registry: Registry,
     temp_dir: TempDir,
 }
@@ -159,12 +159,13 @@ impl FSRegistryTestCase {
         ));
 
         let metadata_store = Arc::new(
-            metadata_store::fs::Backend::new(&metadata_store::fs::BackendConfig {
+            metadata_store::fs::BackendConfig {
                 root_dir: path,
                 sync_to_disk: false,
                 lock_strategy: metadata_store::LockStrategy::Memory,
-            })
-            .unwrap(),
+            }
+            .to_backend()
+            .expect("fs metadata backend"),
         );
         let registry = create_test_registry(
             blob_store.clone(),
@@ -223,7 +224,7 @@ impl RegistryTestCase for FSRegistryTestCase {
 pub struct S3RegistryTestCase {
     key_prefix: String,
     s3_blob_store: Arc<blob_store::s3::Backend>,
-    s3_metadata_store: Arc<metadata_store::s3::Backend>,
+    s3_metadata_store: Arc<metadata_store::Backend>,
     s3_registry: Registry,
 }
 
@@ -248,26 +249,27 @@ impl S3RegistryTestCase {
         );
 
         let metadata_store = Arc::new(
-            metadata_store::s3::Backend::new(
-                &metadata_store::s3::BackendConfig {
-                    access_key_id: Secret::new("root".to_string()),
-                    secret_key: Secret::new("roottoor".to_string()),
-                    endpoint: "http://127.0.0.1:9000".to_string(),
-                    region: "region".to_string(),
-                    bucket: "registry".to_string(),
-                    key_prefix: key_prefix.clone(),
-                    lock_strategy: metadata_store::LockStrategy::Memory,
-                    link_cache_ttl: 0,
-                    access_time_debounce_secs: 0,
-                    capabilities: None,
-                },
+            metadata_store::s3::BackendConfig {
+                access_key_id: Secret::new("root".to_string()),
+                secret_key: Secret::new("roottoor".to_string()),
+                endpoint: "http://127.0.0.1:9000".to_string(),
+                region: "region".to_string(),
+                bucket: "registry".to_string(),
+                key_prefix: key_prefix.clone(),
+                lock_strategy: metadata_store::LockStrategy::Memory,
+                link_cache_ttl: 0,
+                access_time_debounce_secs: 0,
+                capabilities: None,
+            }
+            .to_backend(
                 Some(metadata_store::ConditionalCapabilities {
                     put_if_none_match: true,
                     put_if_match: true,
                     delete_if_match: false,
                 }),
+                None,
             )
-            .unwrap(),
+            .expect("s3 metadata backend"),
         );
 
         let registry = create_test_registry(
